+++ examples/03_native.zig
@@ -45,6 +45,7 @@ fn onWindowEvent(ctx: ?*anyopaque, event: zgpui.platform.WindowEvent) void {
         },
         .a11y_press => {},
         .a11y_adjust => {},
+        .a11y_set_value => {},
     }
 }